requirements.txt
alembic==1.13.3
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker
certifi==2024.8.30
charset-normalizer==3.4.0
click
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==22.0.0
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Mako==1.3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mistune==3.0.2
mysql-connector-python==9.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
yarg==0.1.9
email-validator
psycopg2
psycopg2-binary
langchain>=0.1.0
anthropic>=0.8.0
python-dotenv>=1.0.0
langchain-core
langchain-community
langchain-anthropic
langgraph
textwrap3
pytest
ruff